requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
web3==6.11.3
google-generativeai==0.3.1
aiohttp==3.9.1